Get the setup.py working for pypi. Move the client code into its own subdirectory to make setuptools happy.
