+++ .gitignore
@@ -4,6 +4,7 @@
 # Caches and temp dirs
 /build/
 /dist/
+/client_src/dist/
 *.pyc
 .mypy_cache/
 .cache/
