--- MANIFEST
@@ -1,5 +1,0 @@
-# file GENERATED by distutils, do NOT edit
-client_setup.py
-src/relation_engine_client/__init__.py
-src/relation_engine_client/exceptions.py
-src/relation_engine_client/main.py
