+++ client_src/README.md
@@ -96,3 +96,5 @@ Access the `.resp.text` property on the error object to see the response body fr
 The `raise_not_found` argument was set to `True` and no documents were found in the query.
 
 Access the `.req_body` and `.req_params` properties of the error object to see the request data, or simply print the error to debug.
+
+## Development
