+++ client_src/releng_client.egg-info/PKG-INFO
@@ -1,0 +1,10 @@
+Metadata-Version: 1.0
+Name: releng-client
+Version: 0.0.1
+Summary: KBase Relation Engine API Client Module
+Home-page: https://github.com/kbase/relation_engine_api
+Author: UNKNOWN
+Author-email: UNKNOWN
+License: UNKNOWN
+Description: UNKNOWN
+Platform: UNKNOWN
