+++ client_src/releng_client.egg-info/SOURCES.txt
@@ -1,0 +1,10 @@
+README.md
+setup.py
+relation_engine_client/__init__.py
+relation_engine_client/exceptions.py
+relation_engine_client/main.py
+releng_client.egg-info/PKG-INFO
+releng_client.egg-info/SOURCES.txt
+releng_client.egg-info/dependency_links.txt
+releng_client.egg-info/requires.txt
+releng_client.egg-info/top_level.txt
