+++ client_src/releng_client.egg-info/requires.txt
@@ -1,0 +1,1 @@
+requests>=2
