+++ client_src/test/test_integration.py
@@ -2,8 +2,8 @@ import unittest
 import os
 from uuid import uuid4
 
-from src.relation_engine_client import REClient
-from src.relation_engine_client.exceptions import RERequestError, RENotFound
+from relation_engine_client import REClient
+from relation_engine_client.exceptions import RERequestError, RENotFound
 
 _API_URL = os.environ.get('RE_API_URL', 'http://localhost:5000')
 # See the test schemas here:
